17-okta-pysaml2-django/requirements.txt
Django
django-request-logging
pysaml2
requests